osrd-project: create tailwind preset at root
osrd-project: create base rollup.config generator

osrd-project: add base postcss.config
