+++ ui-core/postcss.config.cjs
@@ -1,12 +1,3 @@
-module.exports = {
-  plugins: {
-    'postcss-import': {},
-    'postcss-assets': {},
-    'tailwindcss/nesting': {},
-    tailwindcss: { config: './tailwind.config.js' },
-    autoprefixer: {},
-    'postcss-preset-env': {
-      features: { 'nesting-rules': false },
-    },
-  },
-};
+const generateBasePostcssConfig = require('../postcss-base.config.cjs');
+
+module.exports = generateBasePostcssConfig();
